Reduced output in vc-client CLI
- Switched from logging to simple eprintlns
- Added --quiet to veracruz-client
- Removed rogue "post_runtime_manager started" println in veracruz_client.rs
- Also updated Shamir shares in example to include newline as a nicety

Ok, so why use eprintln in vc-client but not vc-server or vc-pas?

My reasoning is that because vc-server/vc-pas are both servers, you
would expect their execution to be long-lived with a verbose,
configurable log. But this is not true for vc-client, which is a
short-lived comand used to process single commands at a time.

With vc-server/vc-pas it's helpful to have a verbose log for retroactive
debugging, but this is not the case with vc-client, where a short print
is easily readable/digestable, and further debugging will usually
involve additional commands.

Another note is that the current output of vc-client is roughly modeled
on the output of git fetch: Simple print statements that precede
error-prone/long operations and succeed short operations that may have
more info after completion.